separated tests from components
