+++ apps/client/src/__mocks__/address.js
@@ -1,0 +1,19 @@
+export default {
+  houseNumberFull: "123",
+  restrictions: [
+    {
+      __typename: "Monument",
+      name: "monument",
+    },
+    {
+      __typename: "CityScape",
+      name: "cityscape",
+    },
+  ],
+  streetName: "streetname",
+  zoningPlans: [
+    {
+      name: "zoningplan",
+    },
+  ],
+};
